moved test_list_networks from infra to cloud
